Abstract perception eval operator (#168)
