+++ pylot/perception/base_perception_eval_operator.py
@@ -1,0 +1,201 @@
+import math
+
+import erdos
+
+
+class BasePerceptionEvalOperator(erdos.Operator):
+    """Operator that computes accuracy metrics using tracked obstacles.
+
+    Args:
+        prediction_stream (:py:class:`erdos.ReadStream`): The stream on
+            which tracked obstacles are received.
+        ground_truth_stream: The stream on which
+            :py:class:`~pylot.perception.messages.ObstaclesMessage` are
+            received from the simulator.
+        flags (absl.flags): Object to be used to access absl flags.
+    """
+    def __init__(self, prediction_stream, ground_truth_stream,
+                 finished_indicator_stream, evaluate_timely, matching_policy,
+                 frame_gap, flags):
+        prediction_stream.add_callback(self.on_prediction)
+        ground_truth_stream.add_callback(self.on_ground_truth)
+        erdos.add_watermark_callback([prediction_stream, ground_truth_stream],
+                                     [finished_indicator_stream],
+                                     self.on_watermark)
+        self._logger = erdos.utils.setup_logging(self.config.name,
+                                                 self.config.log_file_name)
+        self._last_notification = None
+        # Buffer of predictions.
+        self._predictions = []
+        # Buffer of ground data.
+        self._ground_truths = []
+        # Storing pairs of (game time, ground/output time).
+        self._prediction_start_end_times = []
+        self._frame_gap = frame_gap
+        self._evaluate_timely = evaluate_timely
+        self._matching_policy = matching_policy
+        # The start time of the most recent inference that completed before
+        # the timestamp of the watermark.
+        self._start_time_best_inference = None
+        # Index in prediction_start_end_times to the inference with the next
+        # unprocessed start time. We need to maintain this index because
+        # the start_tracker_end_times list might contain entries with
+        # start time beyond current watermark.
+        self._start_time_frontier = 0
+        # Buffer storing start times and ground times of predictions we have
+        # to compute accuracy for. The buffer is used to ensure that these
+        # accuracies are computed only once the ground data is available.
+        self._accuracy_compute_buffer = []
+
+    @staticmethod
+    def connect(prediction_stream, ground_truth_stream):
+        """Connects the operator to other streams.
+
+        Args:
+            prediction_stream (:py:class:`erdos.ReadStream`): The stream
+                on which obstacle tracks are received from object trackers.
+            ground_truth_stream: The stream on which
+                :py:class:`~pylot.perception.messages.ObstaclesMessage` are
+                received from the simulator.
+        """
+        finished_indicator_stream = erdos.WriteStream()
+        return [finished_indicator_stream]
+
+    def on_watermark(self, timestamp, finished_indicator_stream):
+        """Invoked when all input streams have received a watermark.
+
+        Args:
+            timestamp (:py:class:`erdos.timestamp.Timestamp`): The timestamp of
+                the watermark.
+        """
+        if timestamp.is_top:
+            return
+        assert len(timestamp.coordinates) == 1
+        game_time = timestamp.coordinates[0]
+        on_new_prediction = False
+        if self._start_time_frontier < len(self._prediction_start_end_times):
+            (st,
+             et) = self._prediction_start_end_times[self._start_time_frontier]
+            if st == game_time:
+                on_new_prediction = True
+                self._accuracy_compute_buffer.append((st, et, False))
+        index = self._start_time_frontier
+        if not on_new_prediction:
+            index = self._start_time_frontier - 1
+        while index >= 0:
+            (p_start_time,
+             p_end_time) = self._prediction_start_end_times[index]
+            if p_end_time == game_time:
+                # This is the result that arrived before start_time, and
+                # uses the most up-to-date sensor data (tracker_start_end_times
+                # is sorted by start_times).
+                self._start_time_best_inference = p_start_time
+                self._prediction_start_end_times = \
+                    self._prediction_start_end_times[index:]
+                self._start_time_frontier -= index
+                break
+            index -= 1
+        if self._start_time_best_inference:
+            # Compute accuracy if we have a prediction with end_time less than
+            # current frontier.
+            self._accuracy_compute_buffer.append(
+                (self._start_time_best_inference, game_time, True))
+        self.__drain_accuracy_compute_buffer(game_time)
+        # Increase the frontier to process the next start time.
+        if on_new_prediction:
+            self._start_time_frontier += 1
+
+    def __drain_accuracy_compute_buffer(self, up_to_time):
+        for (st, et, end_anchored) in self._accuracy_compute_buffer:
+            if et <= up_to_time:
+                self.compute_accuracy(st, et, end_anchored)
+        self._accuracy_compute_buffer = list(
+            filter(lambda args: args[1] > up_to_time,
+                   self._accuracy_compute_buffer))
+        # Can safely GC until the minimum between the min start time in the
+        # accuracy compute buffer and the best inference start time.
+        gc_threshold = min(self._accuracy_compute_buffer, default=[None])[0]
+        if (self._start_time_best_inference is not None
+                and (gc_threshold is None
+                     or gc_threshold > self._start_time_best_inference)):
+            gc_threshold = self._start_time_best_inference
+        if gc_threshold is not None:
+            self.__gc_data_earlier_than(gc_threshold)
+
+    def compute_accuracy(self, frame_time, ground_time, end_anchored):
+        raise NotImplementedError("To be implemented by child class.")
+
+    def __compute_frame_gap(self, game_time):
+        """Infer frame gap if not explicitly provided in constructor."""
+        if not self._frame_gap:
+            if not self._last_notification:
+                self._last_notification = game_time
+                return
+            else:
+                self._frame_gap = (game_time - self._last_notification)
+                self._last_notification = game_time
+
+    def get_ground_truth_at(self, timestamp):
+        for (ground_time, obstacles) in self._ground_truths:
+            if ground_time == timestamp:
+                return obstacles
+            elif ground_time > timestamp:
+                break
+        self._logger.fatal(
+            'Could not find ground obstacles for {}'.format(timestamp))
+
+    def get_prediction_at(self, timestamp):
+        for (start_time, obstacles) in self._predictions:
+            if start_time == timestamp:
+                return obstacles
+            elif start_time > timestamp:
+                break
+        self._logger.fatal(
+            'Could not find tracked obstacles for {}'.format(timestamp))
+
+    def __gc_data_earlier_than(self, game_time):
+        index = 0
+        while (index < len(self._predictions)
+               and self._predictions[index][0] < game_time):
+            index += 1
+        if index > 0:
+            self._predictions = self._predictions[index:]
+        index = 0
+        while (index < len(self._ground_truths)
+               and self._ground_truths[index][0] < game_time):
+            index += 1
+        if index > 0:
+            self._ground_truths = self._ground_truths[index:]
+
+    def on_prediction(self, msg):
+        game_time = msg.timestamp.coordinates[0]
+        self._predictions.append((game_time, msg.obstacles))
+        if len(self._predictions) > 1:
+            assert game_time >= self._predictions[-2][0], \
+                'Obstacle messages did not arrive in order'
+        # Two metrics: 1) mAP, and 2) timely-mAP
+        if not self._evaluate_timely:
+            # We will compare the obstacles with the ground truth at the same
+            # game time.
+            self._prediction_start_end_times.append((game_time, game_time))
+        else:
+            # Ground obstacles time should be as close as possible to the time
+            # of the obstacles + detector + tracker runtime.
+            ground_truth_time = self.__compute_closest_frame_time(game_time +
+                                                                  msg.runtime)
+            self._prediction_start_end_times.append(
+                (game_time, ground_truth_time))
+
+    def on_ground_truth(self, msg):
+        game_time = msg.timestamp.coordinates[0]
+        self._ground_truths.append((game_time, msg.obstacles))
+
+    def __compute_closest_frame_time(self, time):
+        if self._matching_policy == 'ceil':
+            base = math.ceil(time / self._frame_gap) * self._frame_gap
+        elif self._matching_policy == 'round':
+            base = round(time / self._frame_gap) * self._frame_gap
+        else:
+            raise ValueError('Matching policy {} not supported'.format(
+                self._matching_policy))
+        return base
